requirements.txt
gdown
colorama
faiss-cpu
func-timeout
torch
transformers
datasets
tqdm
evaluate
scikit-learn
accelerate
bitsandbytes
protobuf
sentencepiece
pandas
wandb
python-dotenv